Labeler record handling

+++ packages/bsky/src/data-plane/server/indexing/index.ts
@@ -26,6 +26,7 @@ import * as ListItem from './plugins/list-item'
 import * as ListBlock from './plugins/list-block'
 import * as Block from './plugins/block'
 import * as FeedGenerator from './plugins/feed-generator'
+import * as ModService from './plugins/mod-service'
 import RecordProcessor from './processor'
 import { subLogger } from '../../../logger'
 import { retryHttp } from '../../../util/retry'
@@ -44,6 +45,7 @@ export class IndexingService {
     listBlock: ListBlock.PluginType
     block: Block.PluginType
     feedGenerator: FeedGenerator.PluginType
+    modService: ModService.PluginType
   }
 
   constructor(
@@ -63,6 +65,7 @@ export class IndexingService {
       listBlock: ListBlock.makePlugin(this.db, this.background),
       block: Block.makePlugin(this.db, this.background),
       feedGenerator: FeedGenerator.makePlugin(this.db, this.background),
+      modService: ModService.makePlugin(this.db, this.background),
     }
   }
 
@@ -296,6 +299,10 @@ export class IndexingService {
     await this.db.db.deleteFrom('like').where('creator', '=', did).execute()
     await this.db.db
       .deleteFrom('feed_generator')
+      .where('creator', '=', did)
+      .execute()
+    await this.db.db
+      .deleteFrom('mod_service')
       .where('creator', '=', did)
       .execute()
     // lists
